updateOAB.sh
Script will verify that the background auto discover service is enabled, and enable it if it isn't; will verify if the Directory Services server is set to use SSL and user authentication and set those values if they're not. Will also check if the .oab files are present for the user and will download if missing. If the files are present but haven't been updated within the last 24 hours, it will delete those files and download again.